+++ src/renderer/ObjectIdRenderer.cpp
@@ -16,8 +16,8 @@
 
 int ObjectIdRenderer::getObjectId(
     const RenderContext& ctx,
-    double screenPosX,
-    double screenPosY,
+    float screenPosX,
+    float screenPosY,
     Viewport* mainViewport)
 {
     // https://stackoverflow.com/questions/10123601/opengl-read-pixels-from-framebuffer-for-ing-rounded-up-to-255-0xff
@@ -50,8 +50,8 @@ int ObjectIdRenderer::getObjectId(
     const float offsetX = screenW * (vpPos.x + 1.f) / GL_SCREEN_SIZE;
     const float offsetY = screenH * (1.f - (vpPos.y + 1.f) / GL_SCREEN_SIZE);
 
-    const float posx = (screenPosX - offsetX) * ratioX;
-    const float posy = (screenPosY - offsetY) * ratioY;
+    const float posx = (static_cast<float>(screenPosX) - offsetX) * ratioX;
+    const float posy = (static_cast<float>(screenPosY) - offsetY) * ratioY;
 
     if (posx < 0 || posx > w || posy < 0 || posy > h) return -1;
 
@@ -64,7 +64,10 @@ int ObjectIdRenderer::getObjectId(
         //int readFormat;
         //glGetFramebufferParameteriv(GL_FRAMEBUFFER, GL_IMPLEMENTATION_COLOR_READ_FORMAT, &readFormat);
 
-        glReadPixels(posx, m_idBuffer->m_spec.height - posy, 1, 1, GL_RGBA, GL_UNSIGNED_BYTE, data);
+        glReadPixels(
+            static_cast<GLint>(posx),
+            static_cast<GLint>(m_idBuffer->m_spec.height - posy),
+            1, 1, GL_RGBA, GL_UNSIGNED_BYTE, data);
 
         int x = 0;
         m_idBuffer->unbind(ctx);
