+++ src/scene/Scene.h
@@ -85,7 +85,7 @@ public:
     NodeController* getActiveCameraController() const;
 
     void bindComponents(Node* node);
-    int getObjectID(const RenderContext& ctx, double posx, double posy);
+    int getObjectID(const RenderContext& ctx, float posx, float posy);
 
     std::shared_ptr<Viewport> m_mainViewport{ nullptr };
     std::shared_ptr<Viewport> m_rearViewport{ nullptr };
